+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.0-beta.22
 ### Breaking
 * syntax extensions now expect a reference to a function or a macro
+* replace `set-obj!` with `set-object!` [#439](https://github.com/jcubic/lips/issues/439)
 ### Features
 * add debugging helpers (`is-debug`, `set-debug!`, and `inspect`)
 * add `set-hash-syntax!` function [#477](https://github.com/jcubic/lips/issues/477)